Add leave thread to details page, move leaveRoom() to Report
