+++ src/libs/actions/Policy.js
@@ -14,7 +14,6 @@ import * as OptionsListUtils from '../OptionsListUtils';
 import DateUtils from '../DateUtils';
 import * as ReportUtils from '../ReportUtils';
 import Log from '../Log';
-import * as Report from './Report';
 import Permissions from '../Permissions';
 
 const allPolicies = {};
@@ -1147,42 +1146,6 @@ function setWorkspaceInviteMembersDraft(policyID, memberEmails) {
     Onyx.set(`${ONYXKEYS.COLLECTION.WORKSPACE_INVITE_MEMBERS_DRAFT}${policyID}`, memberEmails);
 }
 
-/**
- *
- * @param {String} reportID
- */
-function leaveRoom(reportID) {
-    API.write(
-        'LeaveRoom',
-        {
-            reportID,
-        },
-        {
-            optimisticData: [
-                {
-                    onyxMethod: Onyx.METHOD.SET,
-                    key: `${ONYXKEYS.COLLECTION.REPORT}${reportID}`,
-                    value: {
-                        stateNum: CONST.REPORT.STATE_NUM.SUBMITTED,
-                        statusNum: CONST.REPORT.STATUS.CLOSED,
-                    },
-                },
-            ],
-            failureData: [
-                {
-                    onyxMethod: Onyx.METHOD.SET,
-                    key: `${ONYXKEYS.COLLECTION.REPORT}${reportID}`,
-                    value: {
-                        stateNum: CONST.REPORT.STATE_NUM.OPEN,
-                        statusNum: CONST.REPORT.STATUS.OPEN,
-                    },
-                },
-            ],
-        },
-    );
-    Report.navigateToConciergeChat();
-}
-
 export {
     removeMembers,
     addMembersToWorkspace,
@@ -1212,5 +1175,4 @@ export {
     removeWorkspace,
     setWorkspaceInviteMembersDraft,
     isPolicyOwner,
-    leaveRoom,
 };
